// src/types/MoveInfo.ts
export interface MoveInfo {
    name: string;
    type?: string;
    oldPower?: string;
    newPower?: string;
    oldAccuracy?: string;
    newAccuracy?: string;
    oldType?: string;
    newType?: string;
    oldEffect?: string;
    newEffect?: string;
    oldSecondaryEffectChance?: string;
    newSecondaryEffectChance?: string;
}